backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
passlib[bcrypt]==1.7.*
python-jose[cryptography]==3.3.*
python-multipart==0.0.*
aiofiles==24.*
